Complete delete book feature
