+++ src/components/Book.js
@@ -1,27 +1,28 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-/* eslint-disable */
+
 const Book = props => {
-  const { book, removeBook} = props;
-  
+  const { book, handleRemoveBook } = props;
+
   return (
     <tr>
       <td>{book.id}</td>
       <td>{book.title}</td>
       <td>{book.category}</td>
-      <td><button type="submit" onClick={removeBook}>remove</button></td>
+      <td>
+        <button type="submit" onClick={() => handleRemoveBook(book)}>
+          Remove Book
+        </button>
+      </td>
     </tr>
   );
 };
-/* eslint-enable */
 
 Book.propTypes = {
   book: PropTypes.objectOf(
-    PropTypes.oneOfType([
-      PropTypes.number,
-      PropTypes.string,
-    ]),
+    PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   ).isRequired,
+  handleRemoveBook: PropTypes.func.isRequired,
 };
 
 export default Book;
