+++ src/reducers/book.js
@@ -4,11 +4,8 @@ const booksReducer = (state = [], action) => {
     return newBooks;
   }
   if (action.type === 'REMOVE_BOOK') {
-    const newBooks = state.books.filter(book => book.id !== action.book.id);
-    return [
-      ...state,
-      newBooks,
-    ];
+    const newBooks = state.filter(book => book.id !== action.book.id);
+    return newBooks;
   }
   return state;
 };
